requirements.txt
tencentcloud-sdk-python
mysql-connector-python
pygame
Pillow